requirements.txt
pandas
pillow
wand
autocomplete
